+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+TinyPlugz v0.3.0 [TODO]
+    * [main]  Improve plugin classes look up
+    * [guice] LAZY_SERVICES property can be used to lazily instantiate services returned
+              by getServices.
+    * [guice] INJECTOR_FACTORY property can be specified either as String, Class or Object
+
 TinyPlugz v0.2.0 [16.06.2015]
     * Reworked a lot of internal stuff
     * Lot of bugfixes